Harden callback-url sanitization, public parcel page, and CI pins
- Extract hardened sanitizeCallbackUrl helper into a shared utility and use
  it from the signin page, agreement page, and sign-in client. The previous
  agreement page check only rejected "//" prefixes and missed backslash /
  control-char / double-encoded variants. Tests now import the real function
  instead of duplicating it.
- On /p/* public parcel pages, set Cache-Control: private, no-store and
  Referrer-Policy: no-referrer in middleware so upstream caches don't store
  the PII payload and the unguessable URL isn't leaked via Referer. Add a
  30 req/min/IP rate limit on the page itself as an abuse dampener.
- Pin all GitHub Actions to full commit SHAs (with version comments so
  Dependabot can still update them) to prevent a compromised tag from
  executing arbitrary code in CI.
