+++ .github/workflows/continuous_deployment.yml
@@ -25,10 +25,10 @@ jobs:
 
         steps:
             - name: Checkout repository
-              uses: actions/checkout@v4
+              uses: actions/checkout@34e114876b0b11c390a56381ad16ebd13914f8d5 # v4
 
             - name: Log in to GitHub Container Registry
-              uses: docker/login-action@v3
+              uses: docker/login-action@c94ce9fb468520275223c153574b00df6fe4bcc9 # v3
               with:
                   registry: ghcr.io
                   username: ${{ github.actor }}
@@ -36,7 +36,7 @@ jobs:
 
             - name: Extract Docker metadata (tags, labels)
               id: meta
-              uses: docker/metadata-action@v5
+              uses: docker/metadata-action@c299e40c65443455700f0fdfc63efafe5b349051 # v5
               with:
                   images: ghcr.io/${{ matrix.image_name }}
                   tags: |
@@ -44,10 +44,10 @@ jobs:
                       type=sha,prefix=sha-,format=short
 
             - name: Set up Docker Buildx
-              uses: docker/setup-buildx-action@v3
+              uses: docker/setup-buildx-action@8d2750c68a42422c14e847fe6c8ac0403b4cbd6f # v3
 
             - name: Build and push ${{ matrix.name }}
-              uses: docker/build-push-action@v6
+              uses: docker/build-push-action@10e90e3645eae34f1e60eeb005ba3a3d33f178e8 # v6
               with:
                   context: ${{ matrix.context }}
                   file: ${{ matrix.dockerfile }}
@@ -71,7 +71,7 @@ jobs:
             url: https://staging.matcentralen.com
         steps:
             - name: Deploy to staging
-              uses: appleboy/ssh-action@v1.2.2
+              uses: appleboy/ssh-action@2ead5e36573f08b82fbfce1504f1a4b05a647c6f # v1.2.2
               with:
                   host: ${{ secrets.SERVER_HOST_STAGING }}
                   username: ${{ secrets.SERVER_USER }}
@@ -135,7 +135,7 @@ jobs:
         runs-on: ubuntu-latest
         steps:
             - name: Deploy to production
-              uses: appleboy/ssh-action@v1.2.2
+              uses: appleboy/ssh-action@2ead5e36573f08b82fbfce1504f1a4b05a647c6f # v1.2.2
               with:
                   host: ${{ secrets.SERVER_HOST_PRODUCTION }}
                   username: ${{ secrets.SERVER_USER }}
