+++ .github/workflows/init_deploy.yml
@@ -23,7 +23,7 @@ jobs:
         environment: ${{ github.event.inputs.environment }}
         steps:
             - name: Checkout code
-              uses: actions/checkout@v4
+              uses: actions/checkout@34e114876b0b11c390a56381ad16ebd13914f8d5 # v4
 
             - name: Display target environment
               run: |
@@ -45,7 +45,7 @@ jobs:
                   fi
 
             - name: Deploy via SSH
-              uses: appleboy/ssh-action@v1.2.2
+              uses: appleboy/ssh-action@2ead5e36573f08b82fbfce1504f1a4b05a647c6f # v1.2.2
               with:
                   host: ${{ steps.set-host.outputs.server_host }}
                   username: ${{ secrets.SERVER_USER }}
