+++ app/[locale]/auth/signin/SignInClient.tsx
@@ -6,6 +6,7 @@ import { useState } from "react";
 import { signIn } from "next-auth/react";
 import Image from "next/image";
 import { Github } from "@/components/Icons";
+import { sanitizeCallbackUrl } from "@/app/utils/auth/sanitize-callback-url";
 
 export function SignInClient({
     callbackUrl,
@@ -33,13 +34,9 @@ export function SignInClient({
 
         setIsLoading(true);
         try {
-            // Using absolute path to make sure we don't have pathname prefixing issues
-            const safeCallbackUrl =
-                callbackUrl.startsWith("/") &&
-                !callbackUrl.startsWith("//") &&
-                !callbackUrl.includes("\\")
-                    ? callbackUrl
-                    : "/";
+            // Defense-in-depth: re-sanitize on the client even though the prop
+            // was already sanitized server-side in signin/page.tsx.
+            const safeCallbackUrl = sanitizeCallbackUrl(callbackUrl);
 
             await signIn("github", {
                 callbackUrl: safeCallbackUrl,
